Cleanup duplicate files
- app_main.[ch] was present in both app and zxlib
- remove useless files in App
- rename needed header file
- move functions to be kept
