--- app/src/common/app_main.c
@@ -1,91 +1,0 @@
-/*******************************************************************************
-*   (c) 2018, 2019 Zondax GmbH
-*   (c) 2016 Ledger
-*
-*  Licensed under the Apache License, Version 2.0 (the "License");
-*  you may not use this file except in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*      http://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing, software
-*  distributed under the License is distributed on an "AS IS" BASIS,
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*  See the License for the specific language governing permissions and
-*  limitations under the License.
-********************************************************************************/
-
-#include "app_main.h"
-#include "app_mode.h"
-
-#include <string.h>
-#include <os_io_seproxyhal.h>
-#include <os.h>
-
-#include "view.h"
-#include "actions.h"
-#include "tx.h"
-#include "crypto.h"
-#include "coin.h"
-#include "zxmacros.h"
-
-void extractHDPath(uint32_t rx, uint32_t offset) {
-    if ((rx - offset) < sizeof(uint32_t) * HDPATH_LEN_DEFAULT) {
-        THROW(APDU_CODE_WRONG_LENGTH);
-    }
-
-    MEMCPY(hdPath, G_io_apdu_buffer + offset, sizeof(uint32_t) * HDPATH_LEN_DEFAULT);
-
-    // Check values
-    if (hdPath[0] != HDPATH_0_DEFAULT ||
-        hdPath[1] != HDPATH_1_DEFAULT ||
-        hdPath[3] != HDPATH_3_DEFAULT) {
-        THROW(APDU_CODE_DATA_INVALID);
-    }
-
-    // Limit values unless the app is running in expert mode
-    if (!app_mode_expert()) {
-        for(int i=2; i < HDPATH_LEN_DEFAULT; i++) {
-            // hardened or unhardened values should be below 20
-            if ( (hdPath[i] & 0x7FFFFFFF) > 100) THROW(APDU_CODE_CONDITIONS_NOT_SATISFIED);
-        }
-    }
-}
-
-bool process_chunk(volatile uint32_t *tx, uint32_t rx) {
-    (void) tx;
-
-    const uint8_t payloadType = G_io_apdu_buffer[OFFSET_PAYLOAD_TYPE];
-
-    if (G_io_apdu_buffer[OFFSET_P2] != 0) {
-        THROW(APDU_CODE_INVALIDP1P2);
-    }
-
-    if (rx < OFFSET_DATA) {
-        THROW(APDU_CODE_WRONG_LENGTH);
-    }
-
-    uint32_t added;
-    switch (payloadType) {
-        case 0:
-            tx_initialize();
-            tx_reset();
-            extractHDPath(rx, OFFSET_DATA);
-            return false;
-        case 1:
-            added = tx_append(&(G_io_apdu_buffer[OFFSET_DATA]), rx - OFFSET_DATA);
-            if (added != rx - OFFSET_DATA) {
-                THROW(APDU_CODE_OUTPUT_BUFFER_TOO_SMALL);
-            }
-            return false;
-        case 2:
-            added = tx_append(&(G_io_apdu_buffer[OFFSET_DATA]), rx - OFFSET_DATA);
-            if (added != rx - OFFSET_DATA) {
-                THROW(APDU_CODE_OUTPUT_BUFFER_TOO_SMALL);
-            }
-            return true;
-    }
-
-    THROW(APDU_CODE_INVALIDP1P2);
-}
-
